requirements.txt
Flask==0.10.1
Jinja2==2.7.2
Werkzeug==0.9.4
gunicorn==18.0
python-dateutil==2.2
requests==2.3.0
mysql-python
python-decouple
raven
